+++ test/src/adapter/adapter_test.dart
@@ -18,28 +18,9 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-// It should print doh...
-// ignore_for_file: avoid_print
+import "package:klutter/src/adapter/adapter.dart";
+import "package:test/test.dart";
 
-import "dart:io";
+void main() {
 
-import "package:klutter/klutter.dart";
-
-/// Run tasks for a Producer project.
-Future<void> main(List<String> args) async {
-  print("""
-  ════════════════════════════════════════════
-     KLUTTER (v0.1.2)                               
-  ════════════════════════════════════════════
-  """
-      .ok);
-
-  final pathToRoot = Directory.current.absolutePath;
-  final result = await execute(
-    script: ScriptName.producer,
-    pathToRoot: pathToRoot,
-    arguments: args,
-  );
-
-  print(result);
 }